replaced previous example and updated documentation
