+++ examples/batched-solver/doc/results.dox
@@ -1,16 +1,11 @@
 <h1>Results</h1>
-The following is the expected result:
+The following is the expected result on the reference executor:
 
 @code{.cpp}
 
 Residual norm sqrt(r^T r):
-%%MatrixMarket matrix array real general
-1 1
-2.10788e-15
-%%MatrixMarket matrix array real general
-1 1
-2.10788e-15
-
+ System no. 0: residual norm = 2.16283e-06, internal residual norm = 2.16283e-06, iterations = 6
+ System no. 1: residual norm = 1.07502e-06, internal residual norm = 1.07502e-06, iterations = 6
 @endcode
 
 <h3> Comments about programming and debugging </h3>
