+++ examples/batched-solver/doc/short-intro
@@ -1,1 +1,1 @@
-The simple solver example.
+Using and interfacing with a batched solver.
