+++ examples/batched-solver/doc/tooltip
@@ -1,1 +1,1 @@
-Solve a simple linear system with CG. Read a matrix and right hand side from a file.
+Solve a batch of small linear systems using a batched solver.
